Refactor to give make stream and file writer implementation details private in public ABI
