+++ cpp/src/arrow/ipc/CMakeLists.txt
@@ -30,12 +30,12 @@ set(ARROW_IPC_TEST_LINK_LIBS
 
 set(ARROW_IPC_SRCS
   adapter.cc
-  file.cc
   json.cc
   json-internal.cc
   metadata.cc
   metadata-internal.cc
-  stream.cc
+  reader.cc
+  writer.cc
 )
 
 if(NOT APPLE)
@@ -138,10 +138,11 @@ add_dependencies(arrow_ipc_objlib metadata_fbs)
 # Headers: top level
 install(FILES
   adapter.h
-  file.h
+  api.h
   json.h
   metadata.h
-  stream.h
+  reader.h
+  writer.h
   DESTINATION include/arrow/ipc)
 
 # pkg-config support
@@ -151,3 +152,24 @@ configure_file(arrow-ipc.pc.in
 install(
   FILES "${CMAKE_CURRENT_BINARY_DIR}/arrow-ipc.pc"
   DESTINATION "lib/pkgconfig/")
+
+
+set(UTIL_LINK_LIBS
+  arrow_ipc_static
+  arrow_io_static
+  arrow_static
+  boost_filesystem_static
+  boost_system_static
+  dl)
+
+if (NOT APPLE)
+  set(UTIL_LINK_LIBS
+    ${UTIL_LINK_LIBS}
+    boost_filesystem_static
+    boost_system_static)
+endif()
+
+add_executable(file-to-stream file-to-stream.cc)
+target_link_libraries(file-to-stream ${UTIL_LINK_LIBS})
+add_executable(stream-to-file stream-to-file.cc)
+target_link_libraries(stream-to-file ${UTIL_LINK_LIBS})
